requirements.txt
dotenv
numpy
notebook
pandas
pyarrow
snowflake-snowpark-python
snowflake-snowpark-python[pandas]